Update default gparted live download link to point to i486 iso image
The purpose of the default download link is to make it easier for our
users to quickly download and use gparted live.  This is similar to
the default download file setting in SourceForge.  For both of these
settings, the i486 iso image is the best choice because it runs on the
widest range of hardware.