Add to function page a canonical link for URLs not for gparted.org
Google Webmaster Tools recommend that when the content for two or more
different URLs is similar, a canonical link be used on the
non-preferred pages to point to the preferred page.

We prefer all GParted web site content to be referred to via gparted.org.

See also:
About rel="canonical"
https://support.google.com/webmasters/answer/139394